+++ hetu/impl/kernel/Relu.cu
@@ -11,7 +11,7 @@ __global__ void relu_kernel(const spec_t* input, size_t size, spec_t* output) {
   auto idx = blockIdx.x * blockDim.x + threadIdx.x;
   if (idx >= size)
     return;
-  output[idx] = (input[idx] < 0) ? 0 : input[idx];
+  output[idx] = (double(input[idx]) < 0) ? 0 : input[idx];
 }
 
 template <typename spec_t>
@@ -21,7 +21,7 @@ __global__ void relu_gradient_kernel(const spec_t* input,
   auto idx = blockIdx.x * blockDim.x + threadIdx.x;
   if (idx >= size)
     return;
-  output[idx] = (input[idx] < 0) ? 0 : output_grad[idx];
+  output[idx] = (double(input[idx]) < 0) ? 0 : output_grad[idx];
 }
 
 void ReluCuda(const NDArray& input, NDArray& output, const Stream& stream) {
@@ -68,10 +68,6 @@ void ReluGradientCuda(const NDArray& input, const NDArray& output_grad,
         input->data_ptr<spec_t>(), output_grad->data_ptr<spec_t>(), size,
         input_grad->data_ptr<spec_t>());
     });
-    // CudaStreamSynchronize(cuda_stream);
-    //   HT_LOG_INFO << "_____________up____________\n" << output_grad << 
-    //   "\n" << input_grad
-    //   << "\n__________down_____________";
 }
 
 } // namespace impl
